refactor: :building_construction: Reorganize files structure
