+++ src/modules/destinations/dto/photo-destination.dto.ts
