+++ src/modules/photos/entities/photo.entity.ts
