+++ src/modules/testimonials/testimonials.module.ts
@@ -3,7 +3,7 @@ import { TestimonialsController } from './testimonials.controller';
 import { TestimonialsService } from './testimonials.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Testimonial } from './entities/testimonial.entity';
-import { User } from '../users/entities/user.entity';
+import { User } from '../../modules/users/entities/user.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Testimonial, User])],
