+++ src/modules/testimonials/testimonials.service.ts
@@ -5,7 +5,7 @@ import { Repository } from 'typeorm';
 import { CreateTestimonialDto } from './dto/create-testimonial.dto';
 import { UpdateTestimonialDto } from './dto/update-testimonial.dto';
 import { User } from '../users/entities/user.entity';
-import { Photo } from 'src/photos/entities/photo.entity';
+import { Photo } from '../photos/entities/photo.entity';
 
 @Injectable()
 export class TestimonialsService {
