+++ src/resources/validations/env.validation.ts
